SAM3XA: PAL Open drain fixes. Other minor changes.

+++ os/hal/platforms/SAM3XA/serial_lld.h
@@ -202,9 +202,9 @@ typedef struct {
   /* Config */                                                              \
   const SerialConfig              *config;                                  \
   /* The peripheral id of the UART/USART device */                          \
-  uint32_t                  peripheral_id;                                  \
+  uint8_t                   peripheral_id;                                  \
   /* The IRQ id of the UART/USART device */                                 \
-  uint32_t                  irq_id;                                         \
+  uint8_t                   irq_id;                                         \
   /* if this is an USART or SERIAL device */                                \
   uint8_t                   is_usart;                                       \
   /* Pointer to the device register */                                      \
