+++ os/hal/platforms/SAM3XA/usb_lld.h
@@ -304,12 +304,12 @@ struct USBDriver {
   /**
    * @brief The peripheral id of the UART/USART device
    */
-  uint32_t                  peripheral_id;
+  uint8_t                   peripheral_id;
 
   /**
    * @brief The IRQ id of the UART/USART device
    */
-  uint32_t                  irq_id;
+  uint8_t                   irq_id;
 
   /**
    * @brief Actual endpoint size
